Crate rusoto_dms

AWS Database Migration Service
AWS Database Migration Service (AWS DMS) can migrate your data to and from the most widely used commercial and open-source databases such as Oracle, PostgreSQL, Microsoft SQL Server, Amazon Redshift, MariaDB, Amazon Aurora, MySQL, and SAP Adaptive Server Enterprise (ASE). The service supports homogeneous migrations such as Oracle to Oracle, as well as heterogeneous migrations between different database platforms, such as Oracle to MySQL or SQL Server to PostgreSQL.

For more information about AWS DMS, see What Is AWS Database Migration Service? in the AWS Database Migration User Guide.

If you're using the service, you're probably looking for DatabaseMigrationServiceClient and DatabaseMigrationService.

Structs

AccountQuota
Describes a quota for an AWS account, for example, the number of replication instances allowed.

AddTagsToResourceMessage

AddTagsToResourceResponse

ApplyPendingMaintenanceActionMessage

ApplyPendingMaintenanceActionResponse

AvailabilityZone

Certificate
The SSL certificate that can be used to encrypt connections between the endpoints and the replication instance.

Connection

CreateEndpointMessage

CreateEndpointResponse

CreateEventSubscriptionMessage

CreateEventSubscriptionResponse

CreateReplicationInstanceMessage

CreateReplicationInstanceResponse

CreateReplicationSubnetGroupMessage

CreateReplicationSubnetGroupResponse

CreateReplicationTaskMessage

CreateReplicationTaskResponse

DatabaseMigrationServiceClient
A client for the AWS Database Migration Service API.

DeleteCertificateMessage

DeleteCertificateResponse

DeleteEndpointMessage

DeleteEndpointResponse

DeleteEventSubscriptionMessage

DeleteEventSubscriptionResponse

DeleteReplicationInstanceMessage

DeleteReplicationInstanceResponse

DeleteReplicationSubnetGroupMessage

DeleteReplicationSubnetGroupResponse

DeleteReplicationTaskMessage

DeleteReplicationTaskResponse

DescribeAccountAttributesMessage

DescribeAccountAttributesResponse

DescribeCertificatesMessage

DescribeCertificatesResponse

DescribeConnectionsMessage

DescribeConnectionsResponse

DescribeEndpointTypesMessage

DescribeEndpointTypesResponse

DescribeEndpointsMessage

DescribeEndpointsResponse

DescribeEventCategoriesMessage

DescribeEventCategoriesResponse

DescribeEventSubscriptionsMessage

DescribeEventSubscriptionsResponse

DescribeEventsMessage

DescribeEventsResponse

DescribeOrderableReplicationInstancesMessage

DescribeOrderableReplicationInstancesResponse

DescribePendingMaintenanceActionsMessage

DescribePendingMaintenanceActionsResponse

DescribeRefreshSchemasStatusMessage

DescribeRefreshSchemasStatusResponse

DescribeReplicationInstanceTaskLogsMessage

DescribeReplicationInstanceTaskLogsResponse

DescribeReplicationInstancesMessage

DescribeReplicationInstancesResponse

DescribeReplicationSubnetGroupsMessage

DescribeReplicationSubnetGroupsResponse

DescribeReplicationTaskAssessmentResultsMessage

DescribeReplicationTaskAssessmentResultsResponse

DescribeReplicationTasksMessage

DescribeReplicationTasksResponse

DescribeSchemasMessage

DescribeSchemasResponse

DescribeTableStatisticsMessage

DescribeTableStatisticsResponse

DmsTransferSettings
The settings in JSON format for the DMS Transfer type source endpoint.

DynamoDbSettings

ElasticsearchSettings

Endpoint

Event

EventCategoryGroup

EventSubscription

Filter

ImportCertificateMessage

ImportCertificateResponse

KinesisSettings

ListTagsForResourceMessage

ListTagsForResourceResponse

ModifyEndpointMessage

ModifyEndpointResponse

ModifyEventSubscriptionMessage

ModifyEventSubscriptionResponse

ModifyReplicationInstanceMessage

ModifyReplicationInstanceResponse

ModifyReplicationSubnetGroupMessage

ModifyReplicationSubnetGroupResponse

ModifyReplicationTaskMessage

ModifyReplicationTaskResponse

MongoDbSettings

OrderableReplicationInstance

PendingMaintenanceAction

RebootReplicationInstanceMessage

RebootReplicationInstanceResponse

RedshiftSettings

RefreshSchemasMessage

RefreshSchemasResponse

RefreshSchemasStatus

ReloadTablesMessage

ReloadTablesResponse

RemoveTagsFromResourceMessage

RemoveTagsFromResourceResponse

ReplicationInstance

ReplicationInstanceTaskLog
Contains metadata for a replication instance task log.

ReplicationPendingModifiedValues

ReplicationSubnetGroup

ReplicationTask

ReplicationTaskAssessmentResult
The task assessment report in JSON format.

ReplicationTaskStats

ResourcePendingMaintenanceActions

S3Settings
Settings for exporting data to Amazon S3.

StartReplicationTaskAssessmentMessage

StartReplicationTaskAssessmentResponse

StartReplicationTaskMessage

StartReplicationTaskResponse

StopReplicationTaskMessage

StopReplicationTaskResponse

Subnet

SupportedEndpointType

TableStatistics

TableToReload

Tag

TestConnectionMessage

TestConnectionResponse

VpcSecurityGroupMembership

Enums

AddTagsToResourceError
Errors returned by AddTagsToResource

ApplyPendingMaintenanceActionError
Errors returned by ApplyPendingMaintenanceAction

CreateEndpointError
Errors returned by CreateEndpoint

CreateEventSubscriptionError
Errors returned by CreateEventSubscription

CreateReplicationInstanceError
Errors returned by CreateReplicationInstance

CreateReplicationSubnetGroupError
Errors returned by CreateReplicationSubnetGroup

CreateReplicationTaskError
Errors returned by CreateReplicationTask

DeleteCertificateError
Errors returned by DeleteCertificate

DeleteEndpointError
Errors returned by DeleteEndpoint

DeleteEventSubscriptionError
Errors returned by DeleteEventSubscription

DeleteReplicationInstanceError
Errors returned by DeleteReplicationInstance

DeleteReplicationSubnetGroupError
Errors returned by DeleteReplicationSubnetGroup

DeleteReplicationTaskError
Errors returned by DeleteReplicationTask

DescribeAccountAttributesError
Errors returned by DescribeAccountAttributes

DescribeCertificatesError
Errors returned by DescribeCertificates

DescribeConnectionsError
Errors returned by DescribeConnections

DescribeEndpointTypesError
Errors returned by DescribeEndpointTypes

DescribeEndpointsError
Errors returned by DescribeEndpoints

DescribeEventCategoriesError
Errors returned by DescribeEventCategories

DescribeEventSubscriptionsError
Errors returned by DescribeEventSubscriptions

DescribeEventsError
Errors returned by DescribeEvents

DescribeOrderableReplicationInstancesError
Errors returned by DescribeOrderableReplicationInstances

DescribePendingMaintenanceActionsError
Errors returned by DescribePendingMaintenanceActions

DescribeRefreshSchemasStatusError
Errors returned by DescribeRefreshSchemasStatus

DescribeReplicationInstanceTaskLogsError
Errors returned by DescribeReplicationInstanceTaskLogs

DescribeReplicationInstancesError
Errors returned by DescribeReplicationInstances

DescribeReplicationSubnetGroupsError
Errors returned by DescribeReplicationSubnetGroups

DescribeReplicationTaskAssessmentResultsError
Errors returned by DescribeReplicationTaskAssessmentResults

DescribeReplicationTasksError
Errors returned by DescribeReplicationTasks

DescribeSchemasError
Errors returned by DescribeSchemas

DescribeTableStatisticsError
Errors returned by DescribeTableStatistics

ImportCertificateError
Errors returned by ImportCertificate

ListTagsForResourceError
Errors returned by ListTagsForResource

ModifyEndpointError
Errors returned by ModifyEndpoint

ModifyEventSubscriptionError
Errors returned by ModifyEventSubscription

ModifyReplicationInstanceError
Errors returned by ModifyReplicationInstance

ModifyReplicationSubnetGroupError
Errors returned by ModifyReplicationSubnetGroup

ModifyReplicationTaskError
Errors returned by ModifyReplicationTask

RebootReplicationInstanceError
Errors returned by RebootReplicationInstance

RefreshSchemasError
Errors returned by RefreshSchemas

ReloadTablesError
Errors returned by ReloadTables

RemoveTagsFromResourceError
Errors returned by RemoveTagsFromResource

StartReplicationTaskAssessmentError
Errors returned by StartReplicationTaskAssessment

StartReplicationTaskError
Errors returned by StartReplicationTask

StopReplicationTaskError
Errors returned by StopReplicationTask

TestConnectionError
Errors returned by TestConnection

Traits

DatabaseMigrationService
Trait representing the capabilities of the AWS Database Migration Service API. AWS Database Migration Service clients implement this trait.