Struct rusoto_dms::DatabaseMigrationServiceClient

pub struct DatabaseMigrationServiceClient { /* fields omitted */ }

A client for the AWS Database Migration Service API.

Methods

impl DatabaseMigrationServiceClient

pub fn new(region: Region) -> DatabaseMigrationServiceClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> DatabaseMigrationServiceClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,