Struct rusoto_dms::Endpoint

pub struct Endpoint {
    pub certificate_arn: Option<String>,
    pub database_name: Option<String>,
    pub dynamo_db_settings: Option<DynamoDbSettings>,
    pub endpoint_arn: Option<String>,
    pub endpoint_identifier: Option<String>,
    pub endpoint_type: Option<String>,
    pub engine_name: Option<String>,
    pub external_id: Option<String>,
    pub extra_connection_attributes: Option<String>,
    pub kms_key_id: Option<String>,
    pub mongo_db_settings: Option<MongoDbSettings>,
    pub port: Option<i64>,
    pub s3_settings: Option<S3Settings>,
    pub server_name: Option<String>,
    pub ssl_mode: Option<String>,
    pub status: Option<String>,
    pub username: Option<String>,
}

Fields
certificate_arn: Option<String>

The Amazon Resource Name (ARN) used for SSL connection to the endpoint.

database_name: Option<String>

The name of the database at the endpoint.

dynamo_db_settings: Option<DynamoDbSettings>

The settings for the target DynamoDB database. For more information, see the DynamoDBSettings structure.

endpoint_arn: Option<String>

The Amazon Resource Name (ARN) string that uniquely identifies the endpoint.

endpoint_identifier: Option<String>

The database endpoint identifier. Identifiers must begin with a letter; must contain only ASCII letters, digits, and hyphens; and must not end with a hyphen or contain two consecutive hyphens.

endpoint_type: Option<String>

The type of endpoint.

engine_name: Option<String>

The database engine name. Valid values, depending on the EndPointType, include MYSQL, ORACLE, POSTGRES, MARIADB, AURORA, REDSHIFT, S3, SYBASE, DYNAMODB, MONGODB, and SQLSERVER.

external_id: Option<String>

Value returned by a call to CreateEndpoint that can be used for cross-account validation. Use it on a subsequent call to CreateEndpoint to create the endpoint with a cross-account.

extra_connection_attributes: Option<String>

Additional connection attributes used to connect to the endpoint.

kms_key_id: Option<String>

The KMS key identifier that will be used to encrypt the connection parameters. If you do not specify a value for the KmsKeyId parameter, then AWS DMS will use your default encryption key. AWS KMS creates the default encryption key for your AWS account. Your AWS account has a different default encryption key for each AWS region.

mongo_db_settings: Option<MongoDbSettings>

The settings for the MongoDB source endpoint. For more information, see the MongoDbSettings structure.

port: Option<i64>

The port value used to access the endpoint.

s3_settings: Option<S3Settings>

The settings for the S3 target endpoint. For more information, see the S3Settings structure.

server_name: Option<String>

The name of the server at the endpoint.

ssl_mode: Option<String>

The SSL mode used to connect to the endpoint.

SSL mode can be one of four values: none, require, verify-ca, verify-full.

The default value is none.

status: Option<String>

The status of the endpoint.

username: Option<String>

The user name used to connect to the endpoint.