Struct rusoto_dms::ModifyReplicationInstanceMessage

pub struct ModifyReplicationInstanceMessage {
    pub allocated_storage: Option<i64>,
    pub allow_major_version_upgrade: Option<bool>,
    pub apply_immediately: Option<bool>,
    pub auto_minor_version_upgrade: Option<bool>,
    pub engine_version: Option<String>,
    pub multi_az: Option<bool>,
    pub preferred_maintenance_window: Option<String>,
    pub replication_instance_arn: String,
    pub replication_instance_class: Option<String>,
    pub replication_instance_identifier: Option<String>,
    pub vpc_security_group_ids: Option<Vec<String>>,
}

Fields
allocated_storage: Option<i64>

The amount of storage (in gigabytes) to be allocated for the replication instance.

allow_major_version_upgrade: Option<bool>

Indicates that major version upgrades are allowed. Changing this parameter does not result in an outage and the change is asynchronously applied as soon as possible.

Constraints: This parameter must be set to true when specifying a value for the EngineVersion parameter that is a different major version than the replication instance's current version.

apply_immediately: Option<bool>

Indicates whether the changes should be applied immediately or during the next maintenance window.

auto_minor_version_upgrade: Option<bool>

Indicates that minor version upgrades will be applied automatically to the replication instance during the maintenance window. Changing this parameter does not result in an outage except in the following case and the change is asynchronously applied as soon as possible. An outage will result if this parameter is set to true during the maintenance window, and a newer minor version is available, and AWS DMS has enabled auto patching for that engine version.

engine_version: Option<String>

The engine version number of the replication instance.

multi_az: Option<bool>

Specifies if the replication instance is a Multi-AZ deployment. You cannot set the AvailabilityZone parameter if the Multi-AZ parameter is set to true.

preferred_maintenance_window: Option<String>

The weekly time range (in UTC) during which system maintenance can occur, which might result in an outage. Changing this parameter does not result in an outage, except in the following situation, and the change is asynchronously applied as soon as possible. If moving this window to the current time, there must be at least 30 minutes between the current time and end of the window to ensure pending changes are applied.

Default: Uses existing setting

Format: ddd:hh24:mi-ddd:hh24:mi

Valid Days: Mon | Tue | Wed | Thu | Fri | Sat | Sun

Constraints: Must be at least 30 minutes

replication_instance_arn: String

The Amazon Resource Name (ARN) of the replication instance.

replication_instance_class: Option<String>

The compute and memory capacity of the replication instance.

Valid Values: dms.t2.micro | dms.t2.small | dms.t2.medium | dms.t2.large | dms.c4.large | dms.c4.xlarge | dms.c4.2xlarge | dms.c4.4xlarge

replication_instance_identifier: Option<String>

The replication instance identifier. This parameter is stored as a lowercase string.

vpc_security_group_ids: Option<Vec<String>>

Specifies the VPC security group to be used with the replication instance. The VPC security group must work with the VPC containing the replication instance.