Struct rusoto_dms::MongoDbSettings

pub struct MongoDbSettings {
    pub auth_mechanism: Option<String>,
    pub auth_source: Option<String>,
    pub auth_type: Option<String>,
    pub database_name: Option<String>,
    pub docs_to_investigate: Option<String>,
    pub extract_doc_id: Option<String>,
    pub nesting_level: Option<String>,
    pub password: Option<String>,
    pub port: Option<i64>,
    pub server_name: Option<String>,
    pub username: Option<String>,
}

Fields
auth_mechanism: Option<String>

The authentication mechanism you use to access the MongoDB source endpoint.

Valid values: DEFAULT, MONGODB_CR, SCRAM_SHA_1

DEFAULT – For MongoDB version 2.x, use MONGODB_CR. For MongoDB version 3.x, use SCRAM_SHA_1. This attribute is not used when authType=No.

auth_source: Option<String>

The MongoDB database name. This attribute is not used when authType=NO.

The default is admin.

auth_type: Option<String>

The authentication type you use to access the MongoDB source endpoint.

Valid values: NO, PASSWORD

When NO is selected, user name and password parameters are not used and can be empty.

database_name: Option<String>

The database name on the MongoDB source endpoint.

docs_to_investigate: Option<String>

Indicates the number of documents to preview to determine the document organization. Use this attribute when NestingLevel is set to ONE.

Must be a positive value greater than 0. Default value is 1000.

extract_doc_id: Option<String>

Specifies the document ID. Use this attribute when NestingLevel is set to NONE.

Default value is false.

nesting_level: Option<String>

Specifies either document or table mode.

Valid values: NONE, ONE

Default value is NONE. Specify NONE to use document mode. Specify ONE to use table mode.

password: Option<String>

The password for the user account you use to access the MongoDB source endpoint.

port: Option<i64>

The port value for the MongoDB source endpoint.

server_name: Option<String>

The name of the server on the MongoDB source endpoint.

username: Option<String>

The user name you use to access the MongoDB source endpoint.