Struct rusoto_dms::TableStatistics

pub struct TableStatistics {
    pub ddls: Option<i64>,
    pub deletes: Option<i64>,
    pub full_load_condtnl_chk_failed_rows: Option<i64>,
    pub full_load_error_rows: Option<i64>,
    pub full_load_rows: Option<i64>,
    pub inserts: Option<i64>,
    pub last_update_time: Option<f64>,
    pub schema_name: Option<String>,
    pub table_name: Option<String>,
    pub table_state: Option<String>,
    pub updates: Option<i64>,
    pub validation_failed_records: Option<i64>,
    pub validation_pending_records: Option<i64>,
    pub validation_state: Option<String>,
    pub validation_suspended_records: Option<i64>,
}

Fields
ddls: Option<i64>

The Data Definition Language (DDL) used to build and modify the structure of your tables.

deletes: Option<i64>

The number of delete actions performed on a table.

full_load_condtnl_chk_failed_rows: Option<i64>

The number of rows that failed conditional checks during the Full Load operation (valid only for DynamoDB as a target migrations).

full_load_error_rows: Option<i64>

The number of rows that failed to load during the Full Load operation (valid only for DynamoDB as a target migrations).

full_load_rows: Option<i64>

The number of rows added during the Full Load operation.

inserts: Option<i64>

The number of insert actions performed on a table.

last_update_time: Option<f64>

The last time the table was updated.

schema_name: Option<String>

The schema name.

table_name: Option<String>

The name of the table.

table_state: Option<String>

The state of the tables described.

Valid states: Table does not exist | Before load | Full load | Table completed | Table cancelled | Table error | Table all | Table updates | Table is being reloaded

updates: Option<i64>

The number of update actions performed on a table.

validation_failed_records: Option<i64>

The number of records that failed validation.

validation_pending_records: Option<i64>

The number of records that have yet to be validated.

validation_state: Option<String>

The validation state of the table.

The parameter can have the following values

Not enabled—Validation is not enabled for the table in the migration task.

Pending records—Some records in the table are waiting for validation.

Mismatched records—Some records in the table do not match between the source and target.

Suspended records—Some records in the table could not be validated.

No primary key—The table could not be validated because it had no primary key.

Table error—The table was not validated because it was in an error state and some data was not migrated.

Validated—All rows in the table were validated. If the table is updated, the status can change from Validated.

Error—The table could not be validated because of an unexpected error.

validation_suspended_records: Option<i64>

The number of records that could not be validated.