Trait rusoto_dms::DatabaseMigrationService

pub trait DatabaseMigrationService {
    fn add_tags_to_resource(
        &self, 
        input: AddTagsToResourceMessage
    ) -> RusotoFuture<AddTagsToResourceResponse, AddTagsToResourceError>;

    fn apply_pending_maintenance_action(
        &self, 
        input: ApplyPendingMaintenanceActionMessage
    ) -> RusotoFuture<ApplyPendingMaintenanceActionResponse, ApplyPendingMaintenanceActionError>;

    fn create_endpoint(
        &self, 
        input: CreateEndpointMessage
    ) -> RusotoFuture<CreateEndpointResponse, CreateEndpointError>;

    fn create_event_subscription(
        &self, 
        input: CreateEventSubscriptionMessage
    ) -> RusotoFuture<CreateEventSubscriptionResponse, CreateEventSubscriptionError>;

    fn create_replication_instance(
        &self, 
        input: CreateReplicationInstanceMessage
    ) -> RusotoFuture<CreateReplicationInstanceResponse, CreateReplicationInstanceError>;

    fn create_replication_subnet_group(
        &self, 
        input: CreateReplicationSubnetGroupMessage
    ) -> RusotoFuture<CreateReplicationSubnetGroupResponse, CreateReplicationSubnetGroupError>;

    fn create_replication_task(
        &self, 
        input: CreateReplicationTaskMessage
    ) -> RusotoFuture<CreateReplicationTaskResponse, CreateReplicationTaskError>;

    fn delete_certificate(
        &self, 
        input: DeleteCertificateMessage
    ) -> RusotoFuture<DeleteCertificateResponse, DeleteCertificateError>;

    fn delete_endpoint(
        &self, 
        input: DeleteEndpointMessage
    ) -> RusotoFuture<DeleteEndpointResponse, DeleteEndpointError>;

    fn delete_event_subscription(
        &self, 
        input: DeleteEventSubscriptionMessage
    ) -> RusotoFuture<DeleteEventSubscriptionResponse, DeleteEventSubscriptionError>;

    fn delete_replication_instance(
        &self, 
        input: DeleteReplicationInstanceMessage
    ) -> RusotoFuture<DeleteReplicationInstanceResponse, DeleteReplicationInstanceError>;

    fn delete_replication_subnet_group(
        &self, 
        input: DeleteReplicationSubnetGroupMessage
    ) -> RusotoFuture<DeleteReplicationSubnetGroupResponse, DeleteReplicationSubnetGroupError>;

    fn delete_replication_task(
        &self, 
        input: DeleteReplicationTaskMessage
    ) -> RusotoFuture<DeleteReplicationTaskResponse, DeleteReplicationTaskError>;

    fn describe_account_attributes(
        &self
    ) -> RusotoFuture<DescribeAccountAttributesResponse, DescribeAccountAttributesError>;

    fn describe_certificates(
        &self, 
        input: DescribeCertificatesMessage
    ) -> RusotoFuture<DescribeCertificatesResponse, DescribeCertificatesError>;

    fn describe_connections(
        &self, 
        input: DescribeConnectionsMessage
    ) -> RusotoFuture<DescribeConnectionsResponse, DescribeConnectionsError>;

    fn describe_endpoint_types(
        &self, 
        input: DescribeEndpointTypesMessage
    ) -> RusotoFuture<DescribeEndpointTypesResponse, DescribeEndpointTypesError>;

    fn describe_endpoints(
        &self, 
        input: DescribeEndpointsMessage
    ) -> RusotoFuture<DescribeEndpointsResponse, DescribeEndpointsError>;

    fn describe_event_categories(
        &self, 
        input: DescribeEventCategoriesMessage
    ) -> RusotoFuture<DescribeEventCategoriesResponse, DescribeEventCategoriesError>;

    fn describe_event_subscriptions(
        &self, 
        input: DescribeEventSubscriptionsMessage
    ) -> RusotoFuture<DescribeEventSubscriptionsResponse, DescribeEventSubscriptionsError>;

    fn describe_events(
        &self, 
        input: DescribeEventsMessage
    ) -> RusotoFuture<DescribeEventsResponse, DescribeEventsError>;

    fn describe_orderable_replication_instances(
        &self, 
        input: DescribeOrderableReplicationInstancesMessage
    ) -> RusotoFuture<DescribeOrderableReplicationInstancesResponse, DescribeOrderableReplicationInstancesError>;

    fn describe_pending_maintenance_actions(
        &self, 
        input: DescribePendingMaintenanceActionsMessage
    ) -> RusotoFuture<DescribePendingMaintenanceActionsResponse, DescribePendingMaintenanceActionsError>;

    fn describe_refresh_schemas_status(
        &self, 
        input: DescribeRefreshSchemasStatusMessage
    ) -> RusotoFuture<DescribeRefreshSchemasStatusResponse, DescribeRefreshSchemasStatusError>;

    fn describe_replication_instance_task_logs(
        &self, 
        input: DescribeReplicationInstanceTaskLogsMessage
    ) -> RusotoFuture<DescribeReplicationInstanceTaskLogsResponse, DescribeReplicationInstanceTaskLogsError>;

    fn describe_replication_instances(
        &self, 
        input: DescribeReplicationInstancesMessage
    ) -> RusotoFuture<DescribeReplicationInstancesResponse, DescribeReplicationInstancesError>;

    fn describe_replication_subnet_groups(
        &self, 
        input: DescribeReplicationSubnetGroupsMessage
    ) -> RusotoFuture<DescribeReplicationSubnetGroupsResponse, DescribeReplicationSubnetGroupsError>;

    fn describe_replication_task_assessment_results(
        &self, 
        input: DescribeReplicationTaskAssessmentResultsMessage
    ) -> RusotoFuture<DescribeReplicationTaskAssessmentResultsResponse, DescribeReplicationTaskAssessmentResultsError>;

    fn describe_replication_tasks(
        &self, 
        input: DescribeReplicationTasksMessage
    ) -> RusotoFuture<DescribeReplicationTasksResponse, DescribeReplicationTasksError>;

    fn describe_schemas(
        &self, 
        input: DescribeSchemasMessage
    ) -> RusotoFuture<DescribeSchemasResponse, DescribeSchemasError>;

    fn describe_table_statistics(
        &self, 
        input: DescribeTableStatisticsMessage
    ) -> RusotoFuture<DescribeTableStatisticsResponse, DescribeTableStatisticsError>;

    fn import_certificate(
        &self, 
        input: ImportCertificateMessage
    ) -> RusotoFuture<ImportCertificateResponse, ImportCertificateError>;

    fn list_tags_for_resource(
        &self, 
        input: ListTagsForResourceMessage
    ) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>;

    fn modify_endpoint(
        &self, 
        input: ModifyEndpointMessage
    ) -> RusotoFuture<ModifyEndpointResponse, ModifyEndpointError>;

    fn modify_event_subscription(
        &self, 
        input: ModifyEventSubscriptionMessage
    ) -> RusotoFuture<ModifyEventSubscriptionResponse, ModifyEventSubscriptionError>;

    fn modify_replication_instance(
        &self, 
        input: ModifyReplicationInstanceMessage
    ) -> RusotoFuture<ModifyReplicationInstanceResponse, ModifyReplicationInstanceError>;

    fn modify_replication_subnet_group(
        &self, 
        input: ModifyReplicationSubnetGroupMessage
    ) -> RusotoFuture<ModifyReplicationSubnetGroupResponse, ModifyReplicationSubnetGroupError>;

    fn modify_replication_task(
        &self, 
        input: ModifyReplicationTaskMessage
    ) -> RusotoFuture<ModifyReplicationTaskResponse, ModifyReplicationTaskError>;

    fn reboot_replication_instance(
        &self, 
        input: RebootReplicationInstanceMessage
    ) -> RusotoFuture<RebootReplicationInstanceResponse, RebootReplicationInstanceError>;

    fn refresh_schemas(
        &self, 
        input: RefreshSchemasMessage
    ) -> RusotoFuture<RefreshSchemasResponse, RefreshSchemasError>;

    fn reload_tables(
        &self, 
        input: ReloadTablesMessage
    ) -> RusotoFuture<ReloadTablesResponse, ReloadTablesError>;

    fn remove_tags_from_resource(
        &self, 
        input: RemoveTagsFromResourceMessage
    ) -> RusotoFuture<RemoveTagsFromResourceResponse, RemoveTagsFromResourceError>;

    fn start_replication_task(
        &self, 
        input: StartReplicationTaskMessage
    ) -> RusotoFuture<StartReplicationTaskResponse, StartReplicationTaskError>;

    fn start_replication_task_assessment(
        &self, 
        input: StartReplicationTaskAssessmentMessage
    ) -> RusotoFuture<StartReplicationTaskAssessmentResponse, StartReplicationTaskAssessmentError>;

    fn stop_replication_task(
        &self, 
        input: StopReplicationTaskMessage
    ) -> RusotoFuture<StopReplicationTaskResponse, StopReplicationTaskError>;

    fn test_connection(
        &self, 
        input: TestConnectionMessage
    ) -> RusotoFuture<TestConnectionResponse, TestConnectionError>;
}

Trait representing the capabilities of the AWS Database Migration Service API. AWS Database Migration Service clients implement this trait.

Required methods

fn add_tags_to_resource(
&self,
input: AddTagsToResourceMessage
) -> RusotoFuture<AddTagsToResourceResponse, AddTagsToResourceError>

Adds metadata tags to an AWS DMS resource, including replication instance, endpoint, security group, and migration task. These tags can also be used with cost allocation reporting to track cost associated with DMS resources, or used in a Condition statement in an IAM policy for DMS.

fn apply_pending_maintenance_action(
&self,
input: ApplyPendingMaintenanceActionMessage
) -> RusotoFuture<ApplyPendingMaintenanceActionResponse, ApplyPendingMaintenanceActionError>

Applies a pending maintenance action to a resource (for example, to a replication instance).

fn create_endpoint(
&self,
input: CreateEndpointMessage
) -> RusotoFuture<CreateEndpointResponse, CreateEndpointError>

Creates an endpoint using the provided settings.

fn create_event_subscription(
&self,
input: CreateEventSubscriptionMessage
) -> RusotoFuture<CreateEventSubscriptionResponse, CreateEventSubscriptionError>

Creates an AWS DMS event notification subscription.

You can specify the type of source (SourceType) you want to be notified of, provide a list of AWS DMS source IDs (SourceIds) that triggers the events, and provide a list of event categories (EventCategories) for events you want to be notified of. If you specify both the SourceType and SourceIds, such as SourceType = replication-instance and SourceIdentifier = my-replinstance, you will be notified of all the replication instance events for the specified source. If you specify a SourceType but don't specify a SourceIdentifier, you receive notice of the events for that source type for all your AWS DMS sources. If you don't specify either SourceType nor SourceIdentifier, you will be notified of events generated from all AWS DMS sources belonging to your customer account.

For more information about AWS DMS events, see Working with Events and Notifications in the AWS Database Migration Service User Guide.

fn create_replication_instance(
&self,
input: CreateReplicationInstanceMessage
) -> RusotoFuture<CreateReplicationInstanceResponse, CreateReplicationInstanceError>

Creates the replication instance using the specified parameters.

fn create_replication_subnet_group(
&self,
input: CreateReplicationSubnetGroupMessage
) -> RusotoFuture<CreateReplicationSubnetGroupResponse, CreateReplicationSubnetGroupError>

Creates a replication subnet group given a list of the subnet IDs in a VPC.

fn create_replication_task(
&self,
input: CreateReplicationTaskMessage
) -> RusotoFuture<CreateReplicationTaskResponse, CreateReplicationTaskError>

Creates a replication task using the specified parameters.

fn delete_certificate(
&self,
input: DeleteCertificateMessage
) -> RusotoFuture<DeleteCertificateResponse, DeleteCertificateError>

Deletes the specified certificate.

fn delete_endpoint(
&self,
input: DeleteEndpointMessage
) -> RusotoFuture<DeleteEndpointResponse, DeleteEndpointError>

Deletes the specified endpoint.

All tasks associated with the endpoint must be deleted before you can delete the endpoint.

fn delete_event_subscription(
&self,
input: DeleteEventSubscriptionMessage
) -> RusotoFuture<DeleteEventSubscriptionResponse, DeleteEventSubscriptionError>

Deletes an AWS DMS event subscription.

fn delete_replication_instance(
&self,
input: DeleteReplicationInstanceMessage
) -> RusotoFuture<DeleteReplicationInstanceResponse, DeleteReplicationInstanceError>

Deletes the specified replication instance.

You must delete any migration tasks that are associated with the replication instance before you can delete it.

fn delete_replication_subnet_group(
&self,
input: DeleteReplicationSubnetGroupMessage
) -> RusotoFuture<DeleteReplicationSubnetGroupResponse, DeleteReplicationSubnetGroupError>

Deletes a subnet group.

fn delete_replication_task(
&self,
input: DeleteReplicationTaskMessage
) -> RusotoFuture<DeleteReplicationTaskResponse, DeleteReplicationTaskError>

Deletes the specified replication task.

fn describe_account_attributes(
&self
) -> RusotoFuture<DescribeAccountAttributesResponse, DescribeAccountAttributesError>

Lists all of the AWS DMS attributes for a customer account. The attributes include AWS DMS quotas for the account, such as the number of replication instances allowed. The description for a quota includes the quota name, current usage toward that quota, and the quota's maximum value.

This command does not take any parameters.

fn describe_certificates(
&self,
input: DescribeCertificatesMessage
) -> RusotoFuture<DescribeCertificatesResponse, DescribeCertificatesError>

Provides a description of the certificate.

fn describe_connections(
&self,
input: DescribeConnectionsMessage
) -> RusotoFuture<DescribeConnectionsResponse, DescribeConnectionsError>

Describes the status of the connections that have been made between the replication instance and an endpoint. Connections are created when you test an endpoint.

fn describe_endpoint_types(
&self,
input: DescribeEndpointTypesMessage
) -> RusotoFuture<DescribeEndpointTypesResponse, DescribeEndpointTypesError>

Returns information about the type of endpoints available.

fn describe_endpoints(
&self,
input: DescribeEndpointsMessage
) -> RusotoFuture<DescribeEndpointsResponse, DescribeEndpointsError>

Returns information about the endpoints for your account in the current region.

fn describe_event_categories(
&self,
input: DescribeEventCategoriesMessage
) -> RusotoFuture<DescribeEventCategoriesResponse, DescribeEventCategoriesError>

Lists categories for all event source types, or, if specified, for a specified source type. You can see a list of the event categories and source types in Working with Events and Notifications in the AWS Database Migration Service User Guide.

fn describe_event_subscriptions(
&self,
input: DescribeEventSubscriptionsMessage
) -> RusotoFuture<DescribeEventSubscriptionsResponse, DescribeEventSubscriptionsError>

Lists all the event subscriptions for a customer account. The description of a subscription includes SubscriptionName, SNSTopicARN, CustomerID, SourceType, SourceID, CreationTime, and Status.

If you specify SubscriptionName, this action lists the description for that subscription.

fn describe_events(
&self,
input: DescribeEventsMessage
) -> RusotoFuture<DescribeEventsResponse, DescribeEventsError>

Lists events for a given source identifier and source type. You can also specify a start and end time. For more information on AWS DMS events, see Working with Events and Notifications in the AWS Database Migration User Guide.

fn describe_orderable_replication_instances(
&self,
input: DescribeOrderableReplicationInstancesMessage
) -> RusotoFuture<DescribeOrderableReplicationInstancesResponse, DescribeOrderableReplicationInstancesError>

Returns information about the replication instance types that can be created in the specified region.

fn describe_pending_maintenance_actions(
&self,
input: DescribePendingMaintenanceActionsMessage
) -> RusotoFuture<DescribePendingMaintenanceActionsResponse, DescribePendingMaintenanceActionsError>

For internal use only

fn describe_refresh_schemas_status(
&self,
input: DescribeRefreshSchemasStatusMessage
) -> RusotoFuture<DescribeRefreshSchemasStatusResponse, DescribeRefreshSchemasStatusError>

Returns the status of the RefreshSchemas operation.

fn describe_replication_instance_task_logs(
&self,
input: DescribeReplicationInstanceTaskLogsMessage
) -> RusotoFuture<DescribeReplicationInstanceTaskLogsResponse, DescribeReplicationInstanceTaskLogsError>

Returns information about the task logs for the specified task.

fn describe_replication_instances(
&self,
input: DescribeReplicationInstancesMessage
) -> RusotoFuture<DescribeReplicationInstancesResponse, DescribeReplicationInstancesError>

Returns information about replication instances for your account in the current region.

fn describe_replication_subnet_groups(
&self,
input: DescribeReplicationSubnetGroupsMessage
) -> RusotoFuture<DescribeReplicationSubnetGroupsResponse, DescribeReplicationSubnetGroupsError>

Returns information about the replication subnet groups.

fn describe_replication_task_assessment_results(
&self,
input: DescribeReplicationTaskAssessmentResultsMessage
) -> RusotoFuture<DescribeReplicationTaskAssessmentResultsResponse, DescribeReplicationTaskAssessmentResultsError>

Returns the task assessment results from Amazon S3. This action always returns the latest results.

fn describe_replication_tasks(
&self,
input: DescribeReplicationTasksMessage
) -> RusotoFuture<DescribeReplicationTasksResponse, DescribeReplicationTasksError>

Returns information about replication tasks for your account in the current region.

fn describe_schemas(
&self,
input: DescribeSchemasMessage
) -> RusotoFuture<DescribeSchemasResponse, DescribeSchemasError>

Returns information about the schema for the specified endpoint.

fn describe_table_statistics(
&self,
input: DescribeTableStatisticsMessage
) -> RusotoFuture<DescribeTableStatisticsResponse, DescribeTableStatisticsError>

Returns table statistics on the database migration task, including table name, rows inserted, rows updated, and rows deleted.

Note that the "last updated" column the DMS console only indicates the time that AWS DMS last updated the table statistics record for a table. It does not indicate the time of the last update to the table.

fn import_certificate(
&self,
input: ImportCertificateMessage
) -> RusotoFuture<ImportCertificateResponse, ImportCertificateError>

Uploads the specified certificate.

fn list_tags_for_resource(
&self,
input: ListTagsForResourceMessage
) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>

Lists all tags for an AWS DMS resource.

fn modify_endpoint(
&self,
input: ModifyEndpointMessage
) -> RusotoFuture<ModifyEndpointResponse, ModifyEndpointError>

Modifies the specified endpoint.

fn modify_event_subscription(
&self,
input: ModifyEventSubscriptionMessage
) -> RusotoFuture<ModifyEventSubscriptionResponse, ModifyEventSubscriptionError>

Modifies an existing AWS DMS event notification subscription.

fn modify_replication_instance(
&self,
input: ModifyReplicationInstanceMessage
) -> RusotoFuture<ModifyReplicationInstanceResponse, ModifyReplicationInstanceError>

Modifies the replication instance to apply new settings. You can change one or more parameters by specifying these parameters and the new values in the request.

Some settings are applied during the maintenance window.

fn modify_replication_subnet_group(
&self,
input: ModifyReplicationSubnetGroupMessage
) -> RusotoFuture<ModifyReplicationSubnetGroupResponse, ModifyReplicationSubnetGroupError>

Modifies the settings for the specified replication subnet group.

fn modify_replication_task(
&self,
input: ModifyReplicationTaskMessage
) -> RusotoFuture<ModifyReplicationTaskResponse, ModifyReplicationTaskError>

Modifies the specified replication task.

You can't modify the task endpoints. The task must be stopped before you can modify it.

For more information about AWS DMS tasks, see Working with Migration Tasks in the AWS Database Migration Service User Guide.

fn reboot_replication_instance(
&self,
input: RebootReplicationInstanceMessage
) -> RusotoFuture<RebootReplicationInstanceResponse, RebootReplicationInstanceError>

Reboots a replication instance. Rebooting results in a momentary outage, until the replication instance becomes available again.

fn refresh_schemas(
&self,
input: RefreshSchemasMessage
) -> RusotoFuture<RefreshSchemasResponse, RefreshSchemasError>

Populates the schema for the specified endpoint. This is an asynchronous operation and can take several minutes. You can check the status of this operation by calling the DescribeRefreshSchemasStatus operation.

fn reload_tables(
&self,
input: ReloadTablesMessage
) -> RusotoFuture<ReloadTablesResponse, ReloadTablesError>

Reloads the target database table with the source data.

fn remove_tags_from_resource(
&self,
input: RemoveTagsFromResourceMessage
) -> RusotoFuture<RemoveTagsFromResourceResponse, RemoveTagsFromResourceError>

Removes metadata tags from a DMS resource.

fn start_replication_task(
&self,
input: StartReplicationTaskMessage
) -> RusotoFuture<StartReplicationTaskResponse, StartReplicationTaskError>

Starts the replication task.

For more information about AWS DMS tasks, see Working with Migration Tasks in the AWS Database Migration Service User Guide.

fn start_replication_task_assessment(
&self,
input: StartReplicationTaskAssessmentMessage
) -> RusotoFuture<StartReplicationTaskAssessmentResponse, StartReplicationTaskAssessmentError>

Starts the replication task assessment for unsupported data types in the source database.

fn stop_replication_task(
&self,
input: StopReplicationTaskMessage
) -> RusotoFuture<StopReplicationTaskResponse, StopReplicationTaskError>

Stops the replication task.

fn test_connection(
&self,
input: TestConnectionMessage
) -> RusotoFuture<TestConnectionResponse, TestConnectionError>

Tests the connection between the replication instance and the endpoint.

Loading content...